+++ src/components/BugReport.vue
